Normalize deletecategory function

+++ commands/deleteCategory.js
@@ -31,20 +31,20 @@ module.exports = {
       msg.channel.send(`That course doesn't exist, ${msg.author}`);
       return;
     }
-    if(role) {
-      msg.channel.send(`Deleted channels for ${msg.args[1]}, but category not found, ${msg.author}`);
+    if(role && !existingCategory) {
+      msg.channel.send(`Deleted role for ${msg.args[1]}, but category not found, ${msg.author}`);
 
       await role.delete();
       await deleteRole(role.id);
 
       return;
     }
-    if(existingCategory) {
+    if(existingCategory && !role) {
       msg.channel.send(`Deleted channels for ${msg.args[1]}, but role not found, ${msg.author}`);
 
       await existingCategory.children.forEach((channel) => channel.delete());
       await existingCategory.delete();
-      
+
       return;
     }
